chore: upgrade tree-sitter, tree-sitter-bash, and replace deprecated serde_yaml
Upgrade tree-sitter 0.24→0.26 and tree-sitter-bash 0.23→0.25 for error
recovery improvements, grammar correctness fixes, and continued maintenance.
Replace deprecated serde_yaml with serde_norway, a maintained fork
recommended by RustSec (RUSTSEC-2025-0068). Update all transitive
dependencies to latest compatible versions.

+++ src/parser/convert.rs
@@ -147,7 +147,7 @@ fn convert_list(node: Node, source: &str) -> Statement {
     let mut operators: Vec<ListOp> = Vec::new();
 
     let child_count = node.child_count();
-    for i in 0..child_count {
+    for i in 0..child_count as u32 {
         let child = node.child(i).unwrap();
         if child.is_named() {
             items.push(convert_node(child, source));
